fix: change store to make header global
