+++ src/lib-components/HeaderSmart.vue
@@ -31,13 +31,13 @@ export default {
     },
     computed: {
         primaryMenuItems() {
-            return this.$store.state.headerSmart.header.primary
+            return this.$store.state.header.primary
         },
         secondaryMenuItems() {
-            return this.$store.state.headerSmart.header.secondary
+            return this.$store.state.header.secondary
         },
         isMobile() {
-            return this.$store.state.winWidth <= 1024 ? true : false
+            return this.$store.state.header.winWidth <= 1024 ? true : false
         },
         whichHeader() {
             return this.isMobile ? "header-main-responsive" : "header-main"
